particle can look up the energy group its in

+++ src/tally_class.F90
@@ -70,7 +70,7 @@ contains
 
     class(Tally) :: self
     real(8) :: energy
-    real(8) :: energy_bin
+    integer :: energy_bin
     
     ! Get energy group
     energy_bin = binary_search(self % bins, self % nbins+1, energy)
